transfer mixins into Map.js for testing purposes (#6397)
