+++ spec/suites/layer/marker/MarkerSpec.js
@@ -203,30 +203,22 @@ describe("Marker", function () {
 			expect(icon.hasAttribute('alt')).to.be(false);
 		});
 
-		it("pan map to focus marker", function (done) {
+		it("pan map to focus marker", function () {
 			var marker = L.marker([70, 0], {icon: L.divIcon()});
 			map.addLayer(marker);
 
-			setTimeout(function () {
-				expect(function () {
-					marker._icon.focus();
-				}).to.not.throwException();
-
-				done();
-			}, 100);
+			expect(function () {
+				marker._icon.focus();
+			}).to.not.throwException();
 		});
 
-		it("pan map to focus marker with no iconSize", function (done) {
+		it("pan map to focus marker with no iconSize", function () {
 			var marker = L.marker([70, 0], {icon: L.divIcon({iconSize: null})});
 			map.addLayer(marker);
 
-			setTimeout(function () {
-				expect(function () {
-					marker._panOnFocus();
-				}).to.not.throwException();
-
-				done();
-			}, 100);
+			expect(function () {
+				marker._panOnFocus();
+			}).to.not.throwException();
 		});
 	});
