+++ src/control/Control.js
@@ -1,6 +1,4 @@
-
 import {Class} from '../core/Class';
-import {Map} from '../map/Map';
 import * as Util from '../core/Util';
 import * as DomUtil from '../dom/DomUtil';
 
@@ -114,61 +112,3 @@ export var Control = Class.extend({
 export var control = function (options) {
 	return new Control(options);
 };
-
-/* @section Extension methods
- * @uninheritable
- *
- * Every control should extend from `L.Control` and (re-)implement the following methods.
- *
- * @method onAdd(map: Map): HTMLElement
- * Should return the container DOM element for the control and add listeners on relevant map events. Called on [`control.addTo(map)`](#control-addTo).
- *
- * @method onRemove(map: Map)
- * Optional method. Should contain all clean up code that removes the listeners previously added in [`onAdd`](#control-onadd). Called on [`control.remove()`](#control-remove).
- */
-
-/* @namespace Map
- * @section Methods for Layers and Controls
- */
-Map.include({
-	// @method addControl(control: Control): this
-	// Adds the given control to the map
-	addControl: function (control) {
-		control.addTo(this);
-		return this;
-	},
-
-	// @method removeControl(control: Control): this
-	// Removes the given control from the map
-	removeControl: function (control) {
-		control.remove();
-		return this;
-	},
-
-	_initControlPos: function () {
-		var corners = this._controlCorners = {},
-		    l = 'leaflet-',
-		    container = this._controlContainer =
-		            DomUtil.create('div', l + 'control-container', this._container);
-
-		function createCorner(vSide, hSide) {
-			var className = l + vSide + ' ' + l + hSide;
-
-			corners[vSide + hSide] = DomUtil.create('div', className, container);
-		}
-
-		createCorner('top', 'left');
-		createCorner('top', 'right');
-		createCorner('bottom', 'left');
-		createCorner('bottom', 'right');
-	},
-
-	_clearControlPos: function () {
-		for (var i in this._controlCorners) {
-			DomUtil.remove(this._controlCorners[i]);
-		}
-		DomUtil.remove(this._controlContainer);
-		delete this._controlCorners;
-		delete this._controlContainer;
-	}
-});
