+++ src/layer/vector/index.js
@@ -4,7 +4,6 @@ import {SVG, create, pointsToPath, svg} from './SVG';
 SVG.create = create;
 SVG.pointsToPath = pointsToPath;
 export {SVG, svg};
-import './Renderer.getRenderer';	// This is a bit of a hack, but needed because circular dependencies
 
 export {Path} from './Path';
 export {CircleMarker, circleMarker} from './CircleMarker';
